requirements.txt
Flask
Flask-SocketIO
Pillow
opencv-python-headless==4.5.3
numpy
imutils
gunicorn
eventlet
mediapipe